requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
APScheduler==3.9.1
arrow==1.2.2
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
decorator==4.4.2
execnet==1.9.0
Flask==2.1.3
greenlet==1.1.2
h11==0.13.0
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
moviepy==1.0.3
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
playwright==1.23.1
pluggy==1.0.0
proglog==0.1.10
psycopg2-binary==2.9.4
py==1.11.0
pycparser==2.21
pyee==8.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-base-url==2.0.0
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-ordering==0.6
pytest-playwright==0.3.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
schedule==1.1.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
until==0.1.2
urllib3==1.26.10
websockets==10.1
Werkzeug==2.1.2
wsproto==1.1.0
xlwt==1.3.0
XMind==1.2.0
xmind2testcase==1.5.0
